requirements.txt
asgiref==3.3.1
awsebcli==3.19.3
blinker==1.4
botocore==1.19.63
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
docutils==0.15.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
flask-whooshalchemy3==0.2.0
Flask-WhooshAlchemyPlus==0.7.4
future==0.16.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
linecache2==1.0.0
MarkupSafe==1.1.1
pathspec==0.5.9
pbr==3.1.1
PyJWT==1.7.1
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.5
pywin32==228
PyYAML==5.3.1
requests==2.24.0
semantic-version==2.5.0
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
termcolor==1.1.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.25.11
uuid==1.30
wcwidth==0.1.9
Werkzeug==1.0.1
Whoosh==2.7.4
